feat: 材料費 ceil 公式、體積 .1f 顯示、未達低消補足至 500
- 材料費改為 math.ceil(volume_ml * coefficient)：先乘再取上限
- 體積顯示統一改為小數點後一位（四捨五入）：embed、PDF 明細表、file_details_text 皆適用
- 未達低消接受報價時自動補足至 500：
  - embed 顯示「⚠️ 低消補足 +$XXX（補足至低消 NT$ 500）」欄位
  - PDF 費用表新增「低消補足」條目
  - 接受記錄存 500；拒絕記錄存原始金額（不補足）
- 更新所有相關測試

+++ bot/pricing/engine.py
@@ -43,7 +43,7 @@ class QuoteResult:
 
 def calculate_material_cost(resin: ResinType, volume_ml: float, colored: bool) -> int:
     coefficient = _RESIN_COEFFICIENT[(resin, colored)]
-    return int(math.ceil(volume_ml) * coefficient)
+    return math.ceil(volume_ml * coefficient)
 
 
 def calculate_processing_fee(body_count: int) -> int:
